moved demo app ts into e2e tests
